requirements.txt
flask>=2.0
tensorflow>=2.9
tensorflow-hub
pillow
opencv-python
numpy
